tpnewer/pyproject.toml
[project]
name = "tpnewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Flask",
    "Flask-CORS",
    "python-dotenv",
    "requests"
]
